LPS-58633 refactoring -  multi resources support for TemplateManager
